backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["ilapshin <igorlapshin@list.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12"
Django = "^4.2.6"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
psycopg = "^3.1.12"
djoser = "^2.2.0"
django-cors-headers = "^4.3.0"
djangorestframework-simplejwt = "^5.3.0"
setuptools = "^68.2.2"
unicode-slugify = "^0.1.5"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
